+++ apps/web-antd/src/views/演示使用自行删除/tinymce/index.vue
@@ -15,10 +15,10 @@ const content = ref('');
   <Page title="Tinymce富文本">
     <div class="flex flex-col gap-[16px]">
       <div class="flex items-center gap-[16px]">
-        <span>只读</span>
+        <span>禁用</span>
         <Switch v-model:checked="readonly" />
       </div>
-      <Tinymce v-model="content" :height="800" :options="{ readonly }" />
+      <Tinymce v-model="content" :height="800" :disabled="readonly" />
     </div>
   </Page>
 </template>